JRuby: Fix require not working because of lack of standard libraries.
The org.jruby.jruby maven dependency only gives you a barebones JRuby
interpreter - much of the ruby "standard library" is not available
within the package. jruby-complete gives you the same interpreter but
adds the "standard library" which is packaged inside the JAR.

Obviously this comes at a price (current jruby-complete is about 10mb)
but to do anything proper wrt to Jruby steps I would think it is
necessary to at least have the standard lib.

Could perhaps be worthwhile investigating whether to cut it down a bit,
but for now this fixes the "requires" problem in Issue #3.