caching of the step definitions
This is done because the jruby scripting engine is created once, but world is created multiple times.
The world object is created once per scenario, and so the ruby files need to be reparsed every time to
load in the step definitions into world.

This is quite inefficient, and makes for lots of noise on the jruby side if you've got supporting ruby
files that you're using in your step definitions.

A better solution would be to refactor cucumber-jvm core to have world only be created once.
Another solutino would be to recreate the jruby scripting container each time so that the warnings
are gone. But then we're recreating a jruby scripting container each time and that can make things very slow.