Localise the UndefinedStepTracker in the Runtime
Localise the UndefinedStepTracker in the Runtime and remove that
argument from the constructor. Use that the optionalGlue argument is
null, as the cue to let the Runtime create the default RuntimeGlue.
When the optionalGlue argument is not null, the glue injected with that
argument will be used by the Runtime.