Allow for decreasing thread count in cucumber.runtime.TimeoutTest
The TimeoutTest has failed occationally, also on Travis. One source of
failure was that some extra pool thread existed when the initial thread
count was sampled, therefore the test could fail because the current
thread count was lower than the initial thread count. Now the test
allow the current count to be lower or equal to the initial count.