Make SpringTransactionHooks use the step definition context
Make an ugly hack in the SpringFactory to apply the same spring context
to SpringTransactionHooks as the step definition classes.
Make sure search.feature in examples/spring-txn fails if not executed
in a transaction. Remove @txn annotation from see_messages.feature in
examples/spring-txn, since it cannot run in transaction.