Changed the default output appender to be a PrintStream
Previously the default output was an OutputStreamWriter. Since this is buffered
it resulted in test output being buffered making watching the test output in
real time impossible. Now we use a PrintStream which is the type that the
System.out stream is anyway, while still retaining the fix intoduced to make
the Appender's close() method a no-op.