RuntimeOptions can be instantiated in OSGi containers
RuntimeOptions loads the USAGE.txt on class initialization. However 
gherkin.util.FixJava is in the gherkin jar/bundle and in an OSGi 
container the class loader of FixJava does not see the text file, 
because it is located in the core jar/bundle.

Printing the usage text in an OSGi container is not needed, so the code
has been changed to lazy load the resource.