Allows to configure TestNG to fail on unparseable feature files
TestNG DataProvider did not fail if a feature file was not parseable.
This commit allows to configure whether a test fails if at least one
feature file is not parseable.
In addition, test always fails if at least one feature file is not
parseable and strict mode is enabled.

Fixes cucumber/cucumber-jvm#953