Add the --allow-started-ignored option to the JUnit module.
The --allow-started-ignored option makes the JUnitReporter to always
fire test started before a step is executed, so that the test strated
and test finished notifications can be used to measure the execution
time of the step. In case of pending steps, the test started will be
followed by a test ignored, and that will confuse some listeners.