Handle whitespace in rerun file
If I write a rerun file manually using vi on a Mac and pass it to Cucumber as an option (`@rerun.txt`), it fails to run because it can't find the feature files. Even though the same text in a rerun file generated by Cucumber works.

The error is because there was a newline to the end of the rerun file. The rerun file generated by Cucumber does not have this.

It seems Cucumber shouldn't care whether or not lines have newlines or any other kind of whitespace.