Renamed extension because it was silly.
Wasn't thinking when I named the Junit Jupiter extension
"JunitJupiterExtension". Renamed it to CucumberExtension because
that's sensible.

Also fixed a bug which caused the HTML formatter (and any others
that relied on uri() and done() to be called) to fail to complete
their job. HTML reports are now being generated correctly.