[JUnit] --no-step-notifications, use feature name as class name
Use the feature name as the class name when creating the test
descriptions for the pickles/scenarios/test cases (the leaf of the
description tree). This is consistent with the abstraction presented by
the junit formatter. It also improved the reporting when using Gradle,
so that with the --[no-]step-notifications one can choose to get one
report file per feature or one report file per pickle/scenario/test
case.