Fix race condition in Timeout
We need to ensure a happens before relation exists between these events;
  a. the timer setting the interrupt flag on the execution thread.
  b. terminating and cleaning up the timer
To do this we synchronize on monitor. The atomic boolean is merely a
convenient container.

Additionally Timeout did not throw any exceptions when the callback
was busy waiting but did set the interrupted flag. To resolve this
we check and throw after the completion of the callback.

This fixes #1241