Rename TableConverter to XStreamTableConverter
Also refactored the tests a bit to pave the way
for contract testing of the TableConverter interface.
This will make it easier to gradually build a new
TableConverter implementation that doesn't depend
on XStream.

When we have a new implementation on par with the old
one we should be able to rip out the old one.