Automated deployment to maven central
To reduce the drugery of deploying a release this process will now
be handled by travis. Users will only have to prepare the release.

The deployment process is based on [Deploying to OSSRH with Apache Maven](http://central.sonatype.org/pages/apache-maven.html#deploying-to-ossrh-with-apache-maven-introduction)
and has been automated based on the [Complete guide to continuous deployment to maven central from Travis CI](http://www.debonair.io/post/maven-cd/).

To this end the `oss-parent` pom has been replaced with the
`nexus-staging-maven-plugin`.