[Spring] Add explicit dependency on `jaxb-api`
Java 9 added the Jigsaw module system and moved java.xml.bind into a
module. This module is not active by default. It can be be activated by
running with --add-modules java.xml.bind however as it is scheduled to
be removed in Java 10 it is prudent to migrate away from it now.