fix issue with Glue being shared by multiple threads.
ideally initial glue would be immutable and would then be copied per thread.  this is what i've tried, also tried to reduce the number of big/breaking changes - but would've like to split RuntimeGlue up a bit

JavaBackend seems to be a bit of a culprit, for Java8 lambda support.  Again, I've tried to not introduce too much refactoring/breaking changes, but as a minimum I've had to pass the glue reference into build/dispose world rather than holding a reference to it within the backend.  This is obviously a breaking change to other implementations of Backend outside of this repo, so they should be updated accordingly.