[Core] Support parallel execution of pickles (#1389)
Adds supports for parallel execution of pickles to cucumber-jvm

# Added --threads argument to runtime options

Allows users of the CLI to specify the (max) number of threads to be used to run the tests. TestNG/JUnit users should consult their documentation on how to run JUnit/TestNG in parallel.

Note that JUnit will only run features in parallel, not scenarios.

# Concurrent Events

During parallel executing events from the execution of different pickles may interleave. To avoid breaking existing Formatters, these will now receive all test events after the run is complete. Because we are unable to infer whether JUnit/TestNG run in parallel the assumption is that they are and their formatter will always get all events after the run. 

Formatters that can handle concurrent events can events in real time by implementing the ConcurrentEventListener.

# New formatter introduced TimelineFormatter

Which produces reports using vsjis.org timeline to highlight which feature was run on which Thread and when.