[Core] Parse rerun file in RuntimeOptions
Cucumber runs features on the feature path. The feature path can consist
of features, features with line filters or a rerun file containing
features with line filters.

The current implementation would pass the rerun file onwards as if
it were another feature file. The FeatureLoader would then deal with it.

This behaviour is undesirable for two reasons:

1. It comes as a surprise that the FeatureLoader also parses rerun
   files.
2. It comes as a surprise that the Filters also parses rerun files.
3. The rerun file and syntax are part of the command line runner. Other
   runners have different methods to rerun their own tests. As such it
   should be contained in RuntimeOptions and not leak into other parts
   of the core.

Moving rerun file parsing into CucumberOptions now allows the following
to be done:

1. Focus the the FeatureLoader on loading features only
2. Consolidate parsing resources in to GherkinFeatures and compiling
   GherkinFeature into pickles in a single FeatureParser that now
   produces a ready to use CucumberFeature object.
3. Reduce the responsibility of the FeatureBuilder to building a list of
   features.