[Core] Add logger
To assist debugging and provide more clarity about the internal
state of Cucumber JVM some output is needed. In #147 a `--verbose`
option is suggested but given the integrated nature of
`RuntimeOptions` this is not feasible. Only when Cucumber is used
as a standalone application (i.e. the CLI) it can  assume control
over the logging.

The logger is based on `java.util.Logger`. This logger can be
configured via `java.util.LogManager` or by using the 'JUL to
SLF4J` bridge.