[Core] Formalize feature path
The feature path was poorly defined. It could either be a path or
a path with a `classpath:` prefix. This ambiguity allowed room
for various inconsistencies and misinterpretations.

1. Pickle URI's were not actually uniform. Depending on operating
   system different path seperators would be used.

2. Pickle URI's did not identify a resource. The `classpath:`
   prefix would be dropped after loading a feature file. This
   lead to a fairly complex try-catch-try-again-on-the-classpath
   mechanism when loading rerun file.

By properly defining the feature identifier it becomes possible to
parse it in `RuntimeOptions` and propagate a consistent URI to
other parts of cucumber-jvm. This reduces the need for work
arounds in various places.

Features are identified by a URI. This URI can either be absolute:
`scheme:/absolute/path/to.feature`, or relative to the
current working directory:`scheme:relative/path/to.feature`.

In either form, when the scheme is omitted `file` will be
assumed: `path/to.feature`.

On systems that use a `File.separatorChar` other then `/`
`File.separatorChar` can be used as a path separator. When
doing so when the scheme must be omitted: `path\to.feature`.

It is recommended to use `/` as the path separator.