[Core] Fix illegal argument exception when using root package as glue
While the use of the root package is to be discouraged it is a valid
package and should be regonized as such.

After fixing the parser various problems cropped up because the URI
to packageName and back conversions while passing arguments from
`RuntimeOptions` to `JavaBackend` to `ResourceLoaderClassFinder` did
not account for this. As these conversions were pointless it was
best to remove them.