[Core] Print errors and warnings via logger
Rather then using `System.out` or `System.err` to print errors and
warnings these should be printed via a logger. This allows end users
more control over what is logged when and where.

The exception to this is `RuntimeOptions`. These outputs are part of
the CLI and should always go to `System.out`.