[JUnit] Invoke @BeforeClass before TestRunStarted event
JUnit is running Cucumber so the expectation is that JUnit's
`@BeforeClass` and `@AfterClass` methods are invoked respectively before
and after all events emitted by Cucumber. However because the
`TestRunStarted` event was emitted in the constructor of `Cucumber` it
would precede the the invocation of `@BeforeClass`.

In older versions of Cucumber sending the `TestSourceRead` events was
coupled to reading the test sources. This made it impossible to read the
features ahead of time -required to to create the test description
tree and fail on lexer errors- and send `TestRunStarted` after
`@BeforeClass`.

This is no longer case since #1367 and so this problem can be resolved.