Adding new timeStampMillis field to event and json reports
New timeStampMillis field added to the eventBus as well as the TimeService (and its stub).

This calcualtes the system time elapsed in millis and is later used in the jsonformatter as a time stamp (after conversion).

This field can be used in all the events for timestamp usages if required (unlike nanos this is an accurate representation of the current timestamp), right now its only being used in TestCaseStarted.

I also updated the JUnits accordinly since the date is calculated dynamically. The JSON formatter takes a DateTimeFormatter input to change how the time will be formatted to string format.