[Core] Add TimeStampMillis to TimeStampedEvent
A timestamp is more useful when that timestamp is relative to the epoch.
Unlike `System.currentTimeMillis`, `System.nanoTime` is relative to some
arbitrary point in time.

Additionally the extra granularity has no practical benefit. This is
evidenced by the stats, junit and, testng formatter which all express
the elapsed time in milliseconds.