[Core] Allow runner to register bus as concurrent or serial event source (#1656)
Fixes #1515. Ensures non-ConcurrentEventListener plugins get events immediately when running on a single thread, rather than delayed until all tests are run.

As discussed in #1515 the pretty text output for cli/junit/testNg is also delayed for users running on a single thread from their IDE or commandline. This small change restores immediate text output when running on a single thread, only delaying when multiple threads are involved.

A better fix would be an overhaul of the PrettyFormatter, but that's considerably more work, whereas this would fix a lot of the same use cases.