[Core] Improve hook ordering
While the order in which hooks are added to the glue is undefined there
is the assumption that they are never the less reversed in order. So
given two method both annotated with a before- and after hook, they
should be executed like so:

 - method a as a before hook
 - method b as a before hook
 - method b as a after hook
 - method a as a after hook

This failed to happen because Cucumber sorts hooks by order. This means
that before hooks with lower order are executed before hooks with a
higher order. And in reverse for after hooks. However because sort
preserves relative order when the weights are equal the order depends on
the order the hooks were are added to the glue. As a result we ended up
with this execution pattern:

 - method a as a before hook
 - method b as a before hook
 - method a as a after hook
 - method b as a after hook