[Core] Extract Plugin and Event modules
With the introduction of v5 we will break the plugin API by moving
packages from `cucumber.core` to `io.cucumber.core`. We can use this
as an opportunity to extract the plugin and event classes to separate
modules.

This change moves `io.cucumber.core.plugin` to `io.cucumber.plugin` and
`io.cucumber.core.event` to `io.cucumber.event`. While on its own this
change provides little value it will allow us in the future to provide
better support for plugins.

With the introduction of JPMS we will want to expose these classes for
plugin developers. However we also want to avoid the use of core by most
end users. By moving these classes to their own modules we can avoid the
use of core and future proof the introduction of JPMS.

These classes may have been better of in one single module but I can't
think of a name that would describe such a module accurately.