[Core] Rework backend (#1776)
Introduces the backend module. With this module backend
implementations only need to use dependencies from
`io.cucumber.core.backend`. This will allow us to introduce the
module system later on 

The implementation is not yet perfect. Classpath scanning and
resource loading is located in other modules. Removing this
depends on #1526 and removal of type registry configurer which
depends on #1768 .

Fixes #1386 because steps can now see if an exception came from
user code or actual backend.
Removes timeout which would close #1695 earlier then expected.