[Core] Rename `Mime-Type` to `Media-Type`
Embeddings should be provided with some information to let formatters
and reporters know how to interpret the bytes they receive. This used to
be called `Mime-Type`. However since then as a standard the `Mime-Type`
has been upgraded and renamed to `Media-Type`.

By renaming `Mime-Type` to `Media-Type` we make it clear that consumer
and producers of embeddings should use the semantics outlined in:

https://tools.ietf.org/html/rfc7231#section-3.1.1.1