[JUnit] Support `line` query parameter in `UriSelector` (#1845)
Allows the use of a `line` query parameter when using the `UriSelector`:
  - `classpath:/com/example/example.feature?line=20`
  - `file:/path/to/com/example/example.feature?line=20`

This can be used by IDE creators and users to select a file in a feature file
and easily convert it to something the `CucumberTestEngine` can discover and
execute.

Any `TestDescriptor` that matches the line *and* its descendents will be
included in the discovery result. Note that this differs from the `FeaturePath`
specification. When using `classpath:/com/example/example.feature:20` only
pickles are matched. The query parameter also allows the selection of scenario
(outlines), rules, example sections and whole features.

The the dot-feature-colon-line-number format of `FeaturePath` is intentionally
not reused. As we are using URIs and not encumbered by legacy input formats we
can use the proper abstractions here (additionally the parsing is supported by
the JUnit Platform, which is nice too).