[Spring] Share application context (#1848)
When executing Cucumber in parallel in combination with `cucumber-spring` users
would often encounter a cryptic error message: "No qualifying bean of type
'Steps' available: expected single matching bean but found 1" (#1823).

This was initially fixed by ensuring that the Spring application context was not
shared between threads (#1153, #1148). This however has the downside that the
application context is not shared with JUnit tests either (#1583) and comes at
significant performance penalty (#1846).

The root cause however was a race condition in
`DefaultListableBeanFactory.registerBeanDefinition(String beanName, BeanDefinition beanDefinition)`.
There is a check to see if the definition already exists. If it does not exist,
it replaces the bean definition in a ConcurrentHashMap (in a synchronized block
as well) but then adds the bean name to a list without checking if it already
exists.

By synchronizing inside `SpringFactory.start` we prevent concurrent modification
of the application context by `cucumber-spring` This in turn makes it possible
to share the application context between threads and with JUnit tests.

Fixes: #1846
Fixes: #1583
Closes: #1848
Closes: #1582