[JUnit] Annotate Cucumber with @Testable
@Testable is used to signal to IDEs and tooling vendors that the annotated or
meta-annotated element is testable. This makes it possible to run all feature
files in the package of the annotated class with a single click.

Note that while Cucumber provides support for test discovery through @Testable
the recommended way for IDEs and other tooling use the selectors implemented by
Cucumber to discover tests.