[Archetype] Add Maven Archetype module to project
The maven archetype lives in its own project[1] which makes it
cumbersome to keep it up-to-date with on going releases. As
each release of Cucumber-JVM results in a release of the archetype
it is sensible to keep both project together.

The module uses resource filtering[2] to replace the `{project.version}`
in the archetypes `pom.xml`. This ensures no manual version updates
are needed.

1. https://github.com/cucumber/cucumber-archetype
2. http://maven.apache.org/plugins/maven-resources-plugin