[JUnit] Add module-info for cucumber-junit-platform-engine
Adds module-info for Cucumber Junit Platform Engine implementation.
This allows Java to ensure that all dependencies are available at
boot time.

To ensure backwards compatibility down to Java 8 we are using
multi-release jar files which will contain both Java 8 and
Java 9 class files.

To ensure the right JDK (OpenJDK 11) is used to build the project
the Maven Toolchains plugin will verify that this JDK is used.
This does require adding a `toolchains.xml` file to the ~/.m2/
folder.

Example:

```xml
<?xml version="1.0" encoding="UTF8"?>
<toolchains>
    <toolchain>
        <type>jdk</type>
        <provides>
            <version>11</version>
            <vendor>openjdk</vendor>
        </provides>
        <configuration>
            <jdkHome>/path/to/openjdk11</jdkHome>
        </configuration>
    </toolchain>
</toolchains>
```