[TestNG] Fix NPE when executing empty scenario
The `TestCaseState` defaults to undefined when a scenario has no steps. The
`TestCaseResultListener` however expects there to be an undefined step.

That the `TestCaseState` defaults to undefined is rather unexpected. The status
should be the most severe of the states so far which means that it should
default to passed. Changing this however will require a major release.

So fixing the test case result listeners is the next best thing. It also
appears that there is some duplication ongoing between TestNG and the JUnit
Platform Engine. However this duplication is hard to fix because each
platform uses its own exceptions and TestNG still supports `--non-strict`.

Fixes: https://github.com/cucumber/cucumber-jvm/issues/1885