[Core] TestCaseState should be PASSED by default
Currently `TestCaseState` defaults to `Status.UNDEFINED` when no steps have
been executed. This is odd because the implementation otherwise collects the
most severe result. So the logical expectation would be that it would return
`Status.PASSED`

```java
@Override
public Status getStatus() {
    if (stepResults.isEmpty()) {
        return Status.UNDEFINED;
    }

    Result mostSevereResult = max(stepResults, comparing(Result::getStatus));
    return Status.valueOf(mostSevereResult.getStatus().name());
}
```

As a result the `TestCaseResultObserver` in `junit-platform-engine` and
`testng` needs to check if an undefined scenario was empty or not. This means
our abstraction is rather leaky.

```java
if (status.is(UNDEFINED)) {
    if (suggestions.isEmpty()) {
        // Empty scenario
        return;
     }
    throw new UndefinedStepException(suggestions, strict);
}
```

By stating that the empty test is passed by default and making `TestCaseState`
default to `Status.PASSED` we resolve this complexity. As a result
`Scenario.getState` will return `PASSED` rather then `UNDEFINED` prior to the
execution of the first step of a scenario.