[Core] CLI should search classpath root by default
The Cucumber CLI is rather complex to use. To work correctly it needs both
a `--glue` parameter in the form of a package name and a location of a feature
file in form of a classpath uri or path. As a result people often configure
Cucumber incorrectly and are left wondering why their features or glue can not
be found.

This can be simplified as shown by the Cucumber Platform Engine which defaults
to the classpath root and the JUnit and TestNG runners which default to the
package of the runner class.

So it makes sense to do the same with the CLI.

Assuming the following project layout the CLI can discover and execute
all features without needing additional arguments.

```
├── pom.xml
├── src
│   ├── main
│   │   └── java
│   │       └── com/example/Application.java
│   └── test
│       ├── java
│       │   └── com/example/StepDefinitions.java
│       └── resources
│           └── com/example/example.feature
```

This can be with a single maven command:

```
mvn exec:java                                  \
    -Dexec.classpathScope=test                 \
    -Dexec.mainClass=io.cucumber.core.cli.Main
```

Additional benefits can be found in the fact that IDEA often can't work out
which package should be used as glue and does not provide a default. The
Cucumber Eclipse plugin on the other hand always defaults to the class path
root.

Fixes: https://github.com/cucumber/cucumber-jvm/issues/1875