[JUnit Platform] Support skipping scenarios with cucumber.filter.tags
The JUnit Platform supports skipping tests. For example in JUnit Jupiter a
test can be annotated with `@Disabled`. This test will be marked as skipped.

Cucumber scenarios do not have annotations nor is there any support to disable
specific scenarios. A typical work around with Cucumber JUnit 4 integration was
to set a permanent tag filter for `not @Disabled` in `@CucumberOptions`.

This wasn't possible with the JUnit Platform 5 yet. By adding the
`cucumber.filter.tags` property it becomes possible to recreate JUnit Jupiter's
`@Disabled` functionality.
By setting `cucumber.filter.tags=not @Disabled` property any scenarios tagged
with `@Disabled` will be skipped.

Note that contrary to JUnit 4 skipped scenarios are not removed from the test
hierarchy. This is entirely intentional and a result of the way the JUnit
Platform supports skipping tests.