[All] Remove Jackson from `cucumber-jvm` parent pom (#1900)
The `cucumber-jvm` parent pom included `jackson-databind` in its dependency
management section. Because this is a test only dependency it was marked as
test scoped.

While not intended to function as a bill of materials people were and perhaps
still are using `cucumber-jvm` in this fashion. Removing `jackson-databind`
from the dependency management ensures that `cucumber-jvm` only manages
cucumber and test related dependencies.

While not safe, this does provide a modicum of safety.