[Spring] Do not reuse generic application context between scenarios (#1905)
When neither a glue class with a spring context configuration could be found
nor a `cucumber.xml` file exists `cucumber-spring` will fall back to a
`GenericApplicationContext` to facilitate dependency injection.

The current implementation tries to reuse this application context between
scenarios. This requires refreshing the context which can only be done once.
So it's not possible to reuse the `GenericApplicationContext` and a new one
should be created for each scenario.