[Java] Add @ParameterType(useRegexpMatchAsStrongTypeHint=true/false)
Parameter types are registered with a regex. When a step definition is written
as a regular expression `cucumber-expressions` use the regex in the capture
group as a type hint.

When the type hint provided by the regex and the type hint provided by the
method (i.e. the types of its arguments) disagree about the type
`cucumber-expressions` prefers the hint provided by the regex.

When parameter types are declared with very simple regexes this may cause a
problem. For example:

```
import io.cucumber.java.ParameterType;
import io.cucumber.java.en.Given;

public class Main {

    private static class TypeA {}

    @ParameterType("\\w+")
    public TypeA typeA(String a) {
        return new TypeA();
    }

    @Given("a cucumber expression of {typeA}")
    public void an_object_of_type_a(TypeA a) {
        // works fine!
    }

    @Given("^a regular expression of (\\w+)$")
    public void an_object_of_type_b(String b) {
        // broken: will attempt to use the transformer for TypeA
    }
}
```

By using `@ParameterType(pattern="\\w+", useRegexpMatchAsStrongTypeHint=false)`
this behaviour can be changed to instead prefer the type hint from the method.