[JUnit Platform] Map tags to exclusive resources
Using exclusive resources it is possible to control which scenarios will
not run concurrently with other scenarios that use the same resource.

Cucumber tags can be mapped to exclusive resources. A resource can be\
either locked with a read-write-lock, or a read lock.

For example:

```gherkin
Feature: Exclusive resources

 @my-tag-ab-rw
 Scenario: first example
   Given this reads and writes resource a
   And this reads and writes resource b
   When it is executed it will
   Then it will not be executed concurrently with the second example

 @my-tag-a-r
 Scenario: second example
   Given this reads resource a
   When it is executed it will
   Then it will not be executed concurrently with the first example

```

With this configuration:

```
cucumber.execution.exclusive-resources.my-tag-ab-rw.read-write=resource-a,b
cucumber.execution.exclusive-resources.my-tag-a-r.read=a
```

The first scenario tagged with `@my-tag-ab-rw` will lock resource `a`
and `b` with a read-write lock and will not be concurrently executed with other
scenarios tagged the second scenario that locks resource `a` with a read lock.

Note: The `@` is not included.