[Core] Fix NPE in `CucumberExecutionContext.emitMeta` when in a shaded jar
When used in a shaded jar the manifest files are not present or worse
rewritten. As a result it is not possible to use
`class.getPackage().getImplementationVersion()` to discover the version
of the messages library.

By reading this information from a resource bundle instead we can
ensure that when Cucumber is shaded into another project the
information is preserved.