[Core] Fix line filter for scenario outlines (#1981)
Broken by #1804. While covered by unit tests, the unit tests were broken by
a change in the way feature identifiers were parsed. When parsing a feature
path relative paths are expanded to full file uris. This resulted in the uri
of the pickle not matching the uri provided by the test.

Fixes: #1980