[Core] cucumber.feature preserves tags when used with a feature argument
The feature option can be used with either a rerun file or a feature path. For
example when used as a property:

```
cucumber.feature=path/to/features
cucumber.feature=@path/to/rerun.txt
```

Cucumber options consist of a hierarchy of options. When the feature options is
used with a rerun file; tag and name filters from the previous level should be
removed. This ensures that all scenarios in the rerun file and only the
scenarios in the rerun file should be executed (unless the selection is filtered
more at this or the next level).

When used with a feature argument without a line filter this should not be the case.
However `builder::addRerun` would be invoked anyway.

When parsing `cucumber.feature` for rerun files, when no rerun files are present an
empty collection is returned. Because `splitAndMap` was used rather then
`splitAndThenFlatMap`, `builder::addRerun` would always be invoked.

Fixes: #1985