[JUnit] Make duplicate pickle names unique (#2045)
Individual examples do not have a unique name. And while a unique name is not
required by Junit, various integrations such as sbt-junit assume this the case.

By appending ` [#n]` to a scenario name it becomes unique again. So JUnit 4s
output becomes:

```
A feature with scenario outlines
  A scenario outline #1
  A scenario outline #2
  A scenario outline #3
  A scenario outline #4
```

The `#n` was taken from JUnit 5 which renders examples as:

```
A feature with scenario outlines
  A scenario outline
    With some other text
      Example #1
      Example #2
    With some text
     Example #1
     Example #2
```

Fixes: https://github.com/cucumber/cucumber-jvm-scala/issues/102