[Java8] Allow test execution context to be garbage collected
Glue definitions in Java8 retain a reference to the test execution context.
Glue definitions in turn are used in events. Events in turn may be stored for
the entire duration of Cucumbers execution. So when many scenarios are executed
this may results in significant memory usage.

By removing the reference to the test execution context after the scenario is
executed we can still use step definitions in events without the associated
memory consumption.

Fixes: https://github.com/cucumber/cucumber-jvm/issues/2066