[Core] Fix platform dependent unit tests (#2145)
Ignore `ResourceScannerTest#scanForResourcesPathSymlink` on windows, because to 
make it run windows needs to be in developer mode, and the repo has to be
cloned with actived symlinks.

The `PluginFactoryTest` fails on windows, because the plugins (the real ones,
not the test plugins) which get a file paramter, open the file directly during
instantiation. On windows the test failed, because junit cannot delete the
tmpDir due to open files. The change uses the `EventPublisher` to send events to
the plugin, especially the `TestRunFinishedEvent`. During handling of
the event the plugins close the file and junit can successfully clean up
afterwards.