Ensure that line endings do not get altered by git, otherwise test may fail.
The default configuration of git on windows is to let git do end-of-line conversion, i.e. converting between LF in the repository and CRLF in the working tree. If this happens on the feature files in this directory, tests fail, because the test will read the feature file (with crlf line endings) and compare it to the files describing the expected outcome (containing only lf).