[Compatibility] Fix platform dependent tests (#2148)
* Disabled scanForResourcesDirectorySymlink test on windows, because it only works if the repository was cloned with symlinks activated and windows being in developer mode.

* Add config to ensure that maven runs with -Dfile.encoding=UTF-8. This ensures that the generated sources with UTF-8 filenames like Når.java are correctly handled by maven, when running on operating system with another default encoding (e.g. windows)

* Ensure that line endings do not get altered by git, otherwise test may fail.

The default configuration of git on windows is to let git do end-of-line conversion, i.e. converting between LF in the repository and CRLF in the working tree. If this happens on the feature files in this directory, tests fail, because the test will read the feature file (with crlf line endings) and compare it to the files describing the expected outcome (containing only lf).

Co-authored-by: Kathrin Geilmann <k>