[Core] Improve error message when plugin paths collide
Cucumber v6 introduced a new html formatter. This formatter writes to a file
rather then a directory[1]. However a typical use to ensure all files and up
in a single folder would be `plugin=json:target/cucumber/report.json, html:target/cucumber`.

So depending on the order this results in Cucumber being unable to either
create the `target/cucumber` as file for the html report or `target/cucumber`
as a directory for the json report.

The improved error message should point this out.

Fixes: #2108

Additionally the PluginFactory tests were rewritten to make them follow the
arrange-act-test pattern.

1. https://github.com/cucumber/cucumber-jvm/blob/main/release-notes/v6.0.0.md#improved-html-formatter