[Core] Merge duplicate plugin options (#2191)
It is possible to declare duplicate plugins. Either through `@CucumberOptions`,
or `cucumber.plugin`, ect or a combination thereof. This can be used to ensure
certain plugins are always active during a build in CI.

Merging duplicate plugin options should ensure that plugins are not
instantiated twice with the same output.

Fixes: https://github.com/cucumber/cucumber-jvm/issues/2190