[Core] Synchronize event bus before use (#2358)
We can not exclude the possibility that an event is published to the event bus
after test execution has completed. For example because the test framework may
have decided that a test case has timed out and is now terminating test
execution.

As such events from the pending testcase on the synchronized bus may race with
the events from the unsynchronized bus. This in turn may break the `LinkedList`
in the `CanonicalOrderEventPublisher`.

Fixes: #2288 (mabye??)