[BOM] Add Bill of Materials (#2363)
It is fairly common for one project to use more than one cucumber dependency.
To keep these version in sync a Bill of Materials can be used.

```xml
<dependencyManagement>
    <dependencies>
        <dependency>
            <groupId>io.cucumber</groupId>
            <artifactId>cucumber-bom</artifactId>
            <version>${cucumber.version}</version>
            <type>pom</type>
            <scope>import</scope>
        </dependency>
    </dependencies>
</dependencyManagement>

<!--Then for example-->

<dependencies>
    <dependency>
        <groupId>io.cucumber</groupId>
        <artifactId>cucumber-java</artifactId>
        <scope>test</scope>
    </dependency>
    <dependency>
        <groupId>io.cucumber</groupId>
        <artifactId>cucumber-junit</artifactId>
        <scope>test</scope>
    </dependency>
</dependencies>
```

When using `<scope>import</scope>` all dependencies from the parent pom are
imported. This unfortunately include transitive dependencies from the parent
pom. As such all dependencies from `cucumber-jvm` had to be pushed down to
individual modules.

While this does result in some duplication, it also allows dependencies in
modules to be upgraded independently. This is an acceptable trade off as the
modules already have relatively flew overlapping dependencies. Additionally
with Renovate Bot, the tedium of upgrading is mostly automated.

Closes: #2044