[Core] Fix usage formatter serialization (#2392)
The usage formatter is re(ab)using the json formatter from messages. While it
is technically feasible to come up with a solution that does not do this it
would be better to redesign the formatter with a table-like human-readable
output in mind. As such we can defer this decision to a later point in time.