[Core] Use null safe-messages (#2497)
With https://github.com/cucumber/common/pull/1879 `messages` now uses
`Optional<T>` for optional types avoiding some unexpected null references.
This resulted in some updating of `gherkin` and `html-formatter`.

Additionally with this change the dependency on Jackson has been removed from
`messages`. However we still have to serialize the message objects into json.
So for now Jackson is shaded into `core` instead. And eventually we'll be able
to spin this out to it's own module.