Split changelog per version
We're looking to automate the release process based on
polyglot-release[1]. This requires the use of the changelog tool[2] to
automate various aspects of the process.

Unfortunately the changelog tool puts some requirements on the format
of the changelog. Rather then fixing the formatting on 1k lines it is
preferable to split the changelog and fix only the last 200 lines.

1: https://github.com/cucumber/polyglot-release
2: https://github.com/cucumber/changelog/