Remove the use of the maven-toolchains-plugin
To benefit from GitHubs release automation[1] Cucumber should be able to be
build on an arbitrary jdk docker image. Typically, these do not provide Maven
tool chain configurations.

Instead of going through the effort of setting this up in the
cucumber/action-publish-mvn we can forge the use of toolchains, simplifying
the build process somewhat.

To ensure Cucumber is not build with a JDK version less than 11, the enforcer
plugin will check this now.

Additionally:
 - Allow spotless and revapi to work in sub modules
 - Use surefire and failsafe 3.0.0-M7 everywhere
 - Fix example parent pom versions

1: https://github.com/cucumber/action-publish-mvn