Rename modules and folders to match artifact names (#2596)
Polyglot release[1] assumes that projects follow a certain layout. Specifically
it assumes a project is polyglot if (amongst other conditions) there is a
`java/pom.xml` file.

This project however is not polyglot. The `cucumber-java` module is kep in the
`java` folder. And this trips up Polyglot release. By renaming all folders and
modules to match the artifact names we can avoid this problem.

It also simplifies the project by removing the number of unique identifiers for
(more or less) the same thing. But that is just a nice coincidence.

1: https://github.com/cucumber/polyglot-release