Remove Jackson services from META-INF
Like everyone else Cucumber needs to write Json and does so using Jackson.
To avoid conflicts with other versions of Jackson, Cucumber shades its own
version. Unfortunately this shading isn't done perfectly and the files in
`META-INF/services` are not updated to match.

By excluding the services declared by Jackson we can avoid this problem
altogether, though it may come back in the future if/when Jackson decides
to use its own services.

Fixes: #2620