[Core] Delegate encoding and BOM handling to gherkin (#2624)
Previously before handing over a feature file to the Gherkin parser, Cucumber
would remove any Byte Order Markers (BOM) and determine the encoding of the
feature file based on the `#encoding: <encoding>` comment.

With https://github.com/cucumber/common/pull/2018 this can now be handled by
Gherkin. Removing it simplifies Cucumber a little. Unfortunately the
`FeatureParser` doesn't take a `InputStream` as an argument. So we need to add
a default interface to avoid breaking semver.