[Core] Add faster UUID generator selectable through SPI (#2703)
The cucumber event bus UUID generator can now be configured through a SPI
using a property from:
- command-line
- environment variables
- system properties
- `cucumber.properties`
- `junit-platform.properties`
- `@CucumberOptions`

Two UUID generators are provided:

An event has a UUID. The UUID generator can be configured using the
`cucumber.uuid-generator` property:

- name: io.cucumber.core.eventbus.RandomUuidGenerator*
  features: Thread-safe, collision-free, multi-jvm
  performance (Millions UUID/second):  ~1
  Typical usage example: Reports may be generated on different JVMs at
     the same time. A typical example would be one suite that tests
     against Firefox and another against Safari. The exact browser is
     configured through a property. These are then executed concurrently
     on different Gitlab runners.
- name: io.cucumber.core.eventbus.IncrementingUuidGenerator  
  features: Thread-safe, collision-free, single-jvm
  performance (Millions UUID/second): ~130
  Typical usage example: Reports are generated on a single JVM                                                                                                                                                                                                                                          |

Performance on real projects depends on the size (e.g. on a project with
3095 Given, 445 When, 1177 Then, 445 Scenarios, 48 Rules, the `IncrementingUuidGenerator`
improves the global performance by about 3-5% and the feature file 
parsing performance of about 37%).

The `IncrementingUuidGenerator` is a very simple UUID generator based 
on `AtomicLong` counters. It is thread-safe and collision-free within
a single JVM.

The `RandomUuidGenerator` is the usual `UUID.randomUUID()` generator.

Fixes: #2698