[Pico] Fixed missing calls to start, stop and dispose handles (#2772)
Corrected the pico-container lifecycle: the start, stop and dispose
methods are now called for every test scenarios.

The code is the same as before #2725 (so it's safer in terms of
lifecycle), except that the pico container is recycled in the `start()`
method when it already exists. 

When a second call to `start()` is done (e.g. for the 2nd scenario), 
the container component instances are all in disposed state and the
container lifecycle is in disposed state. The recycling operation
consists in removing all container component instances (i.e. flushing
the cache) and resetting the container lifecycle.


Co-authored-by: Julien Kronegg <julien [at] kronegg.ch>